refactor(router): extract routes/egress.rs (q1 wave 1/5)
Move egress handlers + egress_routes() builder from routes/mod.rs into
a new sibling module routes/egress.rs:

  - egress_learned, egress_learn_toggle, egress_learned_clear
  - egress_fetch, egress_allowlist, egress_pending
  - egress_approve, egress_deny, egress_enforce

Public API unchanged: routes::egress_routes() re-exported via pub use.

Proof of behavioral equivalence:
- 178/178 router tests green (136 + 13 + 26 + 3 = 178, unchanged)
- cargo clippy clean
- item-manifest extractor (syn-based) reports ZERO drift:
  every function body sha256 and every data item preserved byte-exact.
